+++ src/Value/DateAttributes.php
@@ -2,6 +2,10 @@
 
 namespace StadGent\Services\OpeningHours\Value;
 
+use DigipolisGent\Value\ValueAbstract;
+use DigipolisGent\Value\ValueFromArrayInterface;
+use DigipolisGent\Value\ValueInterface;
+
 /**
  * Object describing the date attributes (created & updated) of a record.
  *
@@ -90,7 +94,7 @@ class DateAttributes extends ValueAbstract implements ValueFromArrayInterface
     /**
      * Check if the given value object is the same as this.
      *
-     * @param \StadGent\Services\OpeningHours\Value\ValueInterface $object
+     * @param \DigipolisGent\Value\ValueInterface|\StadGent\Services\OpeningHours\Value\DateAttributes $object
      *
      * @return bool
      */