+++ src/Value/ServiceSource.php
@@ -2,6 +2,10 @@
 
 namespace StadGent\Services\OpeningHours\Value;
 
+use DigipolisGent\Value\ValueAbstract;
+use DigipolisGent\Value\ValueFromArrayInterface;
+use DigipolisGent\Value\ValueInterface;
+
 /**
  * Object describing the service Source.
  *
@@ -82,7 +86,7 @@ class ServiceSource extends ValueAbstract implements ValueFromArrayInterface
     /**
      * Check if the given value object is the same as this.
      *
-     * @param \StadGent\Services\OpeningHours\Value\ValueInterface $object
+     * @param \DigipolisGent\Value\ValueInterface|\StadGent\Services\OpeningHours\Value\ServiceSource $object
      *
      * @return bool
      */